bind statement dialect
